+++ src/gui/item_colordisplay.cpp
@@ -27,11 +27,11 @@ ItemColorDisplay::ItemColorDisplay(Color* color, int id_) :
 
 void
 ItemColorDisplay::draw(DrawingContext& context, const Vector& pos, int menu_width, bool active) {
-  float m = menu_width / 2.f;
-  context.color().draw_filled_rect(Rectf(pos + Vector(16, -8), pos + Vector(m, 8)),
-                                     old_color, 0.0f, LAYER_GUI-1);
-  context.color().draw_filled_rect(Rectf(pos + Vector(m, -8), pos + Vector(menu_width-16, 8)),
-                                     *new_color, 0.0f, LAYER_GUI-1);
+  float m = static_cast<float>(menu_width) / 2.0f;
+  context.color().draw_filled_rect(Rectf(pos + Vector(16, -8), pos + Vector(m, 8.0f)),
+                                   old_color, 0.0f, LAYER_GUI-1);
+  context.color().draw_filled_rect(Rectf(pos + Vector(m, -8), pos + Vector(static_cast<float>(menu_width) - 16.0f, 8.0f)),
+                                   *new_color, 0.0f, LAYER_GUI-1);
 }
 
 int
